[fix] Do the reserved check on the receiving end of the event
The receiving end on should be validating if they are allowed to emit the event. In the previous senario it would be possible that the server emitted an `reconnect` event and it would trigger on the client as it's not blocked on the server. It would also be impossible to emit an `reconnect` event on the server as that event is client is blocked by the client.

This simple trivial change would fix both cases without requiring a list of ALL events that are used in Primus.